Updated usage of SUSYObjDef::initialize.
Was previously using manually set file paths in the initialization,
and not supplying correct arguments for SUSYTools-00-03-19.
I've now removed the file paths and fixed the arguments, with one caveat:
We need to implement a prescription for identifying a sample as MC12b.
This is now an argument to SUSYTools. I've set the arg to false for now,
but this will need to be fixed soon.