Adding MC production flag (MC12a, MC12b)
`m_mcProd` member is added to SusyD3PDAna for specifying the MC production
campaign of an MC sample. This is necessary for the initialization of SUSYTools
because it uses different settings for mc12a and mc12b.

The flag is controlled by the user with the command line argument to NtMaker:
```NtMaker -p mc12a ...```

If a sample is recognized as MC (does not have 'data' in the sample name)
and the MC production flag is *not* set, an error message is printed and the
code aborts in SusyD3PDAna::Begin.