SusyD3PDInterface: add new datamembers
Some of the met componenets are now (i.e. in ntupcommon) read out as
separate objects. I added them to our interface so that we can use
them when calling 'GetMET'. Note that because there are several of
them with the same type, I am now callig 'SetPrefix', but I don't know
whether this is the correct thing to do...does it look right,
@sparticlesteve ?