add susy::D3PDAna class
This class is essentially a copy of SusyD3PDAna, but it inherits
directly from TSelector and is has a D3PDReader::Event datamember
(instead of accessing the branches through an intermediate parent
class SusyD3PDInterface.h). All the branch names used by SUSYTools
have also been updated to the ones available in NTUP_COMMON (except
for a few of them that seem to be missing, see warnings)

There are two reasons for trying this approach:
1. because there's is a class D3PDReader::Event allowing to access all
   the branches, in principle we don't need anymore the container
   class SusyD3PDInterface
2. the class D3PDReader::Event allows to read branches with the same
   type (but different prefixes) in a simpler way because all the
   branches are specified as datamembers

This class compiles fine, now it needs to be tested and debugged.
Then the classes inheriting from SusyD3PDAna will need to be modified
to inherit from D3PDAna, and change the access `d3pd` -> `m_event`.