SusyNtMaker: restore broken leaves
Restore the leaves that I thought were broken (because they didn't
seem to be accessible from the central `D3PDReader`). The pieces of
information are actually there, but I had to re-run the
D3PDReaderMaker (more info on this to come).
The missing leaves were
- isSusySample
- susySpartId1/2
- trueTau
One thing about which I am not sure is whether the change
`trueTauAssocSmall_matched` -> `trueTauAssoc_matched`
@sparticlesteve does is look ok to you?