D3PDAna, SusyNtMaker: switch from Begin to SlaveBegin
For some reason (perhaps the fact that we don't inherit from
SusyD3PDInterface anymore?), Begin was not being called properly
anymore. This would cause the m_isMC flag to be set incorrectly and
all the simulated events to be rejected by the GRL.  It turns out that
TSelector::Begin is deprecated; one should override
TSelector::SlaveBegin instead. See
http://root.cern.ch/root/html/TSelector.html

For now I changed this only in D3PDAna and SusyNtMaker. I will
probably need to make this change to SusyD3PDSkimmer as well when I
drop SusyD3PDAna.