D3PDAna: define d3pdElectrons(), d3pdTaus(), d3pdJets()
Same reasoning as for d3pdMuons(). I believe that these are the only
collections for which one can foresee having several collections for
the same type of objects.

For electrons and jets, there might be some ambiguity to be sorted out
when calling SUSYDef::GetMET(); we will need to investigate this when
we switch from one collection to another...good enough for now.