fix 'Object not connected yet!' issue
Not able to write the output tree because of the error below

Error in <Susy::SusyNtObject::event()>: Object not connected yet!
Error in <D3PDReader::VarHandle::event()>: Failed connecting to D3PD

The error appears also if I have a TTree::Fill() right after
m_susyNt.WriteTo(). This is fixed by changing the include order of
SusyNtObject.h and XaodAnalysis.h. Not clear what the reason is,
but now the output tree is being filled correctly.