the 'bad' auxdata of jets is not always available.
The crash happens for example with this file
```
dq2-get -f "AOD.01534464._000002.pool.root.1" data12_8TeV.00204442.physics_Egamma.merge.AOD.r5723_p1751/
```

This is a hack; the issue seems to be there only for a handful of cases.
Right now I cannot figure out a way to check whether or not the aux data is available.
Going to email the SUSYTools people.

Affects #26