initial version computeHiggsPtUncertaintyParameters
Work in progress. Something is different in our lfv ggh samples.
In the SM ones (for example 160655) I can see a h with status 62
and the number of truth jets is ~ the same as indicated in the
reweigthing instructions.
However, in our lfv samples I cannot see the h with status 62,
and the number of truth jets seems to be larger (before and after
the removal of the ones overlapping with the higgs decay products).
To be investigated.