initial version of configurable python NtMaker
This is mostly a test to get things going. We want to have some way to
configure NtMaker with some configuration file that can have the
necessary parameters for data/mc-bkg/mc-sig.

Initial test:
```
./python/NtMaker.py -d 1 -f DAOD_SUSY1.05247930._000024.pool.root.1

SusyNtMaker::SlaveBegin
getTrigNames -- Trigger names for requested run not available. Exitting.
Warning in <StatusCode>:
Warning in <StatusCode>: Unchecked status codes encountered during the job
Warning in <StatusCode>: Number of unchecked successes: 2

```
Probably something to do with Trigger.cxx on a simulated sample.
Is this a Trigger.cxx bug?