fix the way in which we get the metadata tree
The previous fix (538a3f6) avoided the crash. This change also fixes
the logic. In particular, the way in which one gets a TDirectory from
a TChain is different from the one used for a TTree. Without this change
the directory is missing and input is always classified as non-derived.