Susy::SusyObjId : avoid looping on explicit values, skip eleLooseLH
Rather than having loops with hardcoded begin-end, define functions
returning the list of working points we might want to loop over. This
makes it easier to define all the values but then loop only over a
subset of them.

In particular, SUSYTools-00-07-XX does not allow to initialize the
tool for 'eleLooseLH' (missing scale factors).