SUSYTools configs for el/mu id
Provide a separate SUSYTools configuration file for each of the electron
and muon quality/id working points that we instantiate a SUSYTools
instance for. In the initialization of SUSYTools
(XaodAnalysis::initSusyTools) the loop will now pick out the associated
configuration instead of setting the properties.

Several SUSYTools properties still need to be set by hand: PRW config
and DataSource