Muon trigger SF
right now we build a single-muon xAOD::MuonContainer to feed to the
underlying muon trigger SF tool inside of SUSYTools. This is not
correct. The muon trigger SF tool is intended to provide the SF
corresponding to the group of signal muons that end up in your final
event selection. Here we attach an SF to each individual muon as we are
building an analysis agnostic ntuple.

Will need to consider storing only the efficiencies (data and MC) and
within SusyNtuple calculating the SF on the fly at the analysis level.